+++ magda-typescript-common/src/session/GetUserSession.ts
@@ -1,4 +1,4 @@
-import { Request, Response } from "express";
+import { Request } from "express";
 import { Maybe } from "tsmonad";
 const jwt = require("jsonwebtoken");
 
@@ -15,22 +15,4 @@ export function getUserSession(req: Request, jwtSecret: string): Maybe<any> {
     } else {
         return Maybe.nothing<{}>();
     }
-}
-
-export function getUserGroupsHandling(
-    req: Request,
-    res: Response,
-    jwtSecret: string,
-    cb: (session: any) => void
-) {
-    const session = getUserSession(req, jwtSecret);
-
-    session.caseOf({
-        just: session => {
-            cb(session);
-        },
-        nothing: () => {
-            res.status(200).send("No session info available.");
-        }
-    });
 }